Move snapshot encoders and decoders into a new `Codec` module
